+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Changed
+
+- `mean` and `std deviation` fields are displayed for gas usage while running fuzzing tests 
+
 ## [0.13.1] - 2023-12-20
 
 ### Forge